+++ src/test/typescript/utils.js
@@ -3,64 +3,25 @@ const { writeFileSync } = require('fs');
 const htmlEncoder = require('../../../dist/html-encoder').default;
 
 
-function decodeNode(encodedNode) {
-  return new Function('require', `const exports ={}; ${encodedNode}; return exports;`)(require);
-}
-
-function getOutputString(encodedNode, data) {
-
-  const getNode = decodeNode(encodedNode).getNode;
-  const jsNode = getNode(data);
-
-  return jsNode.toString();
-}
-
 // test() created a JS-based SSR-compatible node and compares it to the expected string
 function test(originalString, data = {}, expectedString, description, testFileName = undefined) {
-  const encodedNode = addTestFixture(htmlEncoder(originalString, 'js', false));
+  const encodedNode = htmlEncoder(originalString, 'code', false);
   try {
-    if (testFileName) {
-      writeFileSync(`${testFileName}.log.js`, encodedNode);
-    }
+    testFileName && writeFileSync(`${testFileName}.log.js`, encodedNode.toString());
 
-    assert.strictEqual(getOutputString(encodedNode, data), expectedString, description);
+    assert.strictEqual(encodedNode.getNode(data).toString(), expectedString, description);
   } catch (err) {
     console.error(err, originalString);
     assert.fail();
   }
-}
-
-// test() created a JS-based SSR-compatible node and compares it to the expected string
-function test(originalString, data = {}, expectedString, description, testFileName = undefined) {
-  const encodedNode = addTestFixture(htmlEncoder(originalString, 'js', false));
-  try {
-    if (testFileName) {
-      writeFileSync(`${testFileName}.log.js`, encodedNode);
-    }
-
-    assert.strictEqual(getOutputString(encodedNode, data), expectedString, description);
-  } catch (err) {
-    console.error(err, originalString);
-    assert.fail();
-  }
-}
-
-function addTestFixture(encodedNode) {
-  return `
-  const xmldom_1 = require("@xmldom/xmldom");
-  const window = { DOMParser: xmldom_1.DOMParser };
-  ${encodedNode}
-  `;
 }
 
 function getNodeFactory(htmlString, testFileName = undefined) {
-  const encoded = addTestFixture(htmlEncoder(htmlString));
+  const encoded = htmlEncoder(htmlString, 'code');
 
-  if (testFileName) {
-    writeFileSync(`${testFileName}.log.js`, encoded);
-  }
+  testFileName && writeFileSync(`${testFileName}.log.js`, encoded.toString());
 
-  return decodeNode(encoded).getNode;
+  return encoded.getNode;
 }
 
 // getNode() returns a JS-based browser-compatible node (we forcefully inject DOMParser in so it could still run server-side)
@@ -69,12 +30,11 @@ function getNode(htmlString, data, testFileName = undefined) {
 }
 
 function getSsrHtml(htmlString, data, testFileName = undefined) {
-  const ssrEncoded = htmlEncoder(htmlString, 'js', true);
+  const ssrEncoded = htmlEncoder(htmlString, 'code', true);
 
-  testFileName && writeFileSync(`${testFileName}.server.js`, ssrEncoded);
+  testFileName && writeFileSync(`${testFileName}.server.js`, ssrEncoded.toString());
 
-  const getNode = decodeNode(ssrEncoded).getNode;
-  const node = getNode(data);
+  const node = ssrEncoded.getNode(data);
 
   testFileName && writeFileSync(`${testFileName}.html`, node.toString());
 
@@ -84,16 +44,11 @@ function getSsrHtml(htmlString, data, testFileName = undefined) {
 // getSSRNode() creates a JS-based SSR-compatible node but then use it as a base for browser-compatible node to return
 function getSSRNode(htmlString, data, testFileName = undefined) {
   const ssrEncodedNode = getSsrHtml(htmlString, data, testFileName);
-  const browserEncoded = `
-  const xmldom_1 = require("@xmldom/xmldom");
-  const window = { DOMParser: xmldom_1.DOMParser };
-  ${htmlEncoder(htmlString, 'js', false)}
-  `;
+  const browserEncoded = htmlEncoder(htmlString, 'code', false);
 
-  testFileName && writeFileSync(`${testFileName}.browser.js`, browserEncoded);
+  testFileName && writeFileSync(`${testFileName}.browser.js`, browserEncoded.toString());
 
-  const initNode = decodeNode(browserEncoded).initNode;
-  return initNode(ssrEncodedNode);
+  return browserEncoded.initNode(ssrEncodedNode);
 }
 
 function getTSString(htmlString, testFileName = undefined) {